requirements.local.txt
fastapi
sqlmodel
python-dotenv
uvicorn
pytest
httpx
pytest-cov
google-generativeai